+++ astroquery/splatalogue/build_species_table.py
@@ -9,21 +9,51 @@ import json
 import os
 import requests
 
+from astropy.config import paths
 
-def data_path(filename):
+
+def data_path(filename: str):
+    """
+    Build the path to save a file.  Note that this path is part of the
+    astroquery source code, not the astropy cache directory, as the existence
+    of the file is a prerequisite for performing queries.
+
+    Parameters
+    ----------
+    filename : str
+        Name of the file (generally should be splat-species.json)
+
+    Returns
+    -------
+    str
+        Full path to the cache directory
+    """
     data_dir = os.path.join(os.path.dirname(__file__), 'data')
     return os.path.join(data_dir, filename)
 
 
-def get_json_species_ids(outfile='species.json'):
+def get_json_species_ids(outfile='splat-species.json'):
     """
-    Load the NRAO Splatalogue form and parse the inputs into a JSON object
+    Uses BeautifulSoup to scrape the NRAO Splatalogue species
+    selector form, and caches the result as JSON. The file
+    is saved to the ``astropy`` cache.
+
+    Parameters
+    ----------
+    outfile : str, optional
+        Name of the output JSON, by default 'splat-species.json'
+
+    Returns
+    -------
+    str
+        Formatted string representation of the JSON object
     """
     import bs4
 
     result = requests.get('https://www.cv.nrao.edu/php/splat/b.php')
     page = bs4.BeautifulSoup(result.content, 'html5lib')
-    sid = page.findAll('select', attrs={'id': 'sid'})[0]
+    # The ID needs to be checked periodically if Splatalogue is updated
+    sid = page.findAll('select', attrs={'id': 'speciesselectbox'})[0]
 
     species_types = set()
     for kid in sid.children: